runner: fix `stop` race condition
the race condition was as follows: user stops server -> within 7s, user starts server -> after 7s, thread started by `stop` route sees that server is still running and kills the server

this was because the `stop` route used to create a task that killed the server if it was still running after 7 seconds.

to fix this, the `server_observer` now does not set the `server_stopping` bool, a task spawned by `stop` route checks if the server is running over a period of 5 seconds, at 500ms intervals.